backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.2.0
django-filter==23.5
razorpay==1.4.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9